增加套餐商品列表

+++ app/src/main/java/com/shian/shianlife/mvp/pay/view/IWeChatPrePayView.java
@@ -12,5 +12,5 @@ public interface IWeChatPrePayView extends BaseMVPView {
 
     void wechatPrePayFail(String msg);
 
-    Float wechatGetPayTotal();
+    Integer wechatGetPayTotal();
 }
